Fix httpd ordering, add rspec test
The ordering is wrong here, as the docroot tries to depend on the 503
file that inside the docroot.  I missed this because I was on a live
system where the file existed.  Move to a more standard "www_base"
like in other modules.

Add a deployment test to find things like this.

Change-Id: I9b048836fa9942f021f81f7a744bcc9af96a10be